Tidy Frequency Tables

cp_tidyTable returns a frequency table with counts and percentages of values from a character, integer, double, or logical atomic vector

cp_tidyTable(dataVector, sortByCount = FALSE, round = 1, missRow = TRUE, totalRow = TRUE)

Arguments

dataVector
Character, integer, double, or logical atomic vector

sortByCount
Logical scalar that determines if output will be sorted by count or name

round
Number of decimal places displayed

missRow
Logical scalar that determines if the NA row is displayed

totalRow
Logical scalar that determines if the output will have a summary row appended

Source

frequencies package (source code released under GPL-3 license)

Value

a tibble containing the counts and percentages of each value from the provided data

Examples

tableData <- data.frame(
   id = c(1, 2, 3, 4, 5, 6),
   nhoodStr = c("Patch", "Bevo Mill", "Bevo Mill", "Lindenwood Park", "Carondelet" , "Shaw"),
   visit = c(TRUE, FALSE, FALSE, TRUE, TRUE, NA),
   type = as.factor(c("A", "B", "B", "A", "B", "B")),
   stringsAsFactors = FALSE
   )

cp_tidyTable(tableData$nhoodStr)

#> # A tibble: 6 x 4
#>   data            Count Percentage Cum. 
#>   <chr>           <chr> <chr>      <chr>
#> 1 Bevo Mill       2     33.3       33.3 
#> 2 Carondelet      1     16.7       50.0 
#> 3 Lindenwood Park 1     16.7       66.7 
#> 4 Patch           1     16.7       83.4 
#> 5 Shaw            1     16.7       100.0
#> 6 Total           6     100.0      ""   

cp_tidyTable(tableData$nhoodStr, round = 2)

#> # A tibble: 6 x 4
#>   data            Count Percentage Cum.  
#>   <chr>           <chr> <chr>      <chr> 
#> 1 Bevo Mill       2     33.33      33.33 
#> 2 Carondelet      1     16.67      50.00 
#> 3 Lindenwood Park 1     16.67      66.67 
#> 4 Patch           1     16.67      83.34 
#> 5 Shaw            1     16.67      100.00
#> 6 Total           6     100.00     ""    

cp_tidyTable(tableData$visit, round = 2, missRow = TRUE)

#> # A tibble: 4 x 4
#>   data  Count Percentage Cum.  
#>   <chr> <chr> <chr>      <chr> 
#> 1 FALSE 2     33.33      33.33 
#> 2 TRUE  3     50.00      83.33 
#> 3 <NA>  1     16.67      100.00
#> 4 Total 6     100.00     ""    

cp_tidyTable(tableData$visit, round = 2, missRow = FALSE)

#> # A tibble: 3 x 4
#>   data  Count Percentage Cum.  
#>   <chr> <chr> <chr>      <chr> 
#> 1 FALSE 2     40.00      40.00 
#> 2 TRUE  3     60.00      100.00
#> 3 Total 5     100.00     ""    

cp_tidyTable(tableData$type, round = 3, sortByCount = FALSE)

#> # A tibble: 3 x 4
#>   data  Count Percentage Cum.   
#>   <chr> <chr> <chr>      <chr>  
#> 1 A     2     33.333     33.333 
#> 2 B     4     66.667     100.000
#> 3 Total 6     100.000    ""     

cp_tidyTable(tableData$type, round = 3, sortByCount = TRUE)

#> # A tibble: 3 x 4
#>   data  Count Percentage Cum.   
#>   <chr> <chr> <chr>      <chr>  
#> 1 B     4     66.667     66.667 
#> 2 A     2     33.333     100.000
#> 3 Total 6     100.000    ""